lambda/requirements.txt
aws-lambda-powertools==1.25.10
mysql-connector-python==8.0.27
